requirements.txt
pinecone-client
python-dotenv
tinydb
requests
groq
discord.py
